Refactoring replaceing several print calls with log calls.
